requirements.txt
Django==4.1
psycopg2-binary==2.9.3
django-environ==0.9.0
django-phonenumber-field[phonenumberslite]==6.3.0